moved interfaces to shared typings
